+++ core/src/com/tann/dice/gameplay/entity/Monster.java
@@ -20,7 +20,7 @@ public class Monster extends DiceEntity {
 
 
     @Override
-    public float getPixelSize() {
+    public int getPixelSize() {
         return size.pixels;
     }
